tests/eval/pyproject.toml
[project]
name = "xatu-mcp-eval"
version = "0.1.0"
description = "LLM evaluation harness for xatu-mcp MCP server"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "deepeval>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "openai>=1.0.0",  # For OpenRouter evaluator support
    "langfuse>=2.0.0",  # LLM observability and trace viewing
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
xatu-eval = "scripts.run_eval:main"
xatu-repl = "scripts.repl:main"
xatu-langfuse = "scripts.langfuse:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "agent", "metrics", "cases", "scripts"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
